requirements.txt
transformers>=4.30.0
torch>=2.0.0
Pillow>=9.0.0
accelerate>=0.20.0
einops>=0.6.0
numpy>=1.20.0
sentencepiece>=0.1.90
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0 